+++ storage/ontap/ontap_nas.go
@@ -19,7 +19,17 @@ type OntapNASStorageDriver struct {
 func (d *OntapNASStorageDriver) GetStorageBackendSpecs(backend *storage.StorageBackend) error {
 
 	backend.Name = "ontapnas_" + d.Config.DataLIF
-	return getStorageBackendSpecsCommon(d, backend)
+	poolAttrs := d.GetStoragePoolAttributes()
+	return getStorageBackendSpecsCommon(d, backend, poolAttrs)
+}
+
+func (d *OntapNASStorageDriver) GetStoragePoolAttributes() map[string]sa.Offer {
+
+	return map[string]sa.Offer{
+		sa.BackendType:      sa.NewStringOffer(d.Name()),
+		sa.Snapshots:        sa.NewBoolOffer(true),
+		sa.ProvisioningType: sa.NewStringOffer("thick", "thin"),
+	}
 }
 
 func (d *OntapNASStorageDriver) GetVolumeOpts(
@@ -31,10 +41,7 @@ func (d *OntapNASStorageDriver) GetVolumeOpts(
 }
 
 func (d *OntapNASStorageDriver) GetInternalVolumeName(name string) string {
-	return getInternalVolumeNameCommon(
-		storage.GetCommonInternalVolumeName(d.Config.CommonStorageDriverConfig,
-			name),
-	)
+	return getInternalVolumeNameCommon(d.Config.CommonStorageDriverConfig, name)
 }
 
 func (d *OntapNASStorageDriver) CreatePrepare(